Add new Enforcement Action page
Currently, we use the Document Detail page type to create pages about
Enforcement Actions in Wagtail, and use the Related Metadata streamfield
block to store important information about the action. This means that
the metadata (which should be present on all enforcement actions) might
be missing or malformed, and is not easily searchable or filterable.

This PR adds an Enforcement Action page type, which includes a new tab
in the wagtail interface called "Metadata". It contains database fields
for the required metadata. This will allow us to filter on these values
in the Enforcement Actions filtered list page in a future update.